System.Posix.IO.ByteString: lots of new functions and documentation
BUG: We're trying to use Cabal-style MIN_VERSION_foo(1,2,3) macros
in order to offer backwards compatibility with older versions of
the unix package. But, hsc2hs doesn't like those macros and gives
syntax errors. That's okay because we don't /need/ hsc2hs functionality
in this file--- but, with CPP we get an error from our includes on
OSX 10.5.8 about "architecture not supported" (even though the
includes worked fine with hsc2hs before we added the Cabal macros).